service/requirements.txt
# Rest API
Django==4.2.11
djangorestframework==3.15.1
python-slugify==8.0.4

# Security
django-cors-headers==4.3.1
django-environ==0.11.2

## Auth
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

## WebSockets
channels==4.1.0
channels-redis==4.2.0
uvicorn==0.29.0
rsa==4.9 # Optional
LWE4==0.0.0.2 # Optional

## DDBB
psycopg2==2.9.9 # PostgreSQL
django-redis==5.4.0

## Files
pillow==10.3.0
whitenoise==6.6.0
# botocore
# boto3

# Apache Kafka
#confluent-kafka==2.3.0